Process Phase SQL Code
During this phase I used SQL the following to create the avg_ride_length_seconds and ride_count columns to detail the ride duration for Cyclistic customers and the number of rides for each day of the week. This helped to distinguish the number of rides and duration for annual versus casual customers for each day of the week over the past year.