Add files via upload
The screenshot file shows the results of running the query for one of the SQL files. Running the query cleaned and organized the data into just 14 columns of information that is easily readable. The details of the columns show the average number of rides and the average duration over 7-day periods for each month for annual and casual customers.